+++ agent/structs/config_entry_gateways.go
@@ -916,6 +916,34 @@ type BoundAPIGatewayConfigEntry struct {
 	RaftIndex
 }
 
+func (e *BoundAPIGatewayConfigEntry) IsSame(other *BoundAPIGatewayConfigEntry) bool {
+	listeners := map[string]BoundAPIGatewayListener{}
+	for _, listener := range e.Listeners {
+		listeners[listener.Name] = listener
+	}
+
+	otherListeners := map[string]BoundAPIGatewayListener{}
+	for _, listener := range other.Listeners {
+		otherListeners[listener.Name] = listener
+	}
+
+	if len(listeners) != len(otherListeners) {
+		return false
+	}
+
+	for name, listener := range listeners {
+		otherListener, found := otherListeners[name]
+		if !found {
+			return false
+		}
+		if !listener.IsSame(otherListener) {
+			return false
+		}
+	}
+
+	return true
+}
+
 // IsInitializedForGateway returns whether or not this bound api gateway is initialized with the given api gateway
 // including having corresponding listener entries for the gateway.
 func (e *BoundAPIGatewayConfigEntry) IsInitializedForGateway(gateway *APIGatewayConfigEntry) bool {
@@ -1059,10 +1087,6 @@ func (l BoundAPIGatewayListener) IsSame(other BoundAPIGatewayListener) bool {
 // and protocol. Be sure to check both of these before attempting
 // to bind a route to the listener.
 func (l *BoundAPIGatewayListener) BindRoute(routeRef ResourceReference) bool {
-	if l == nil {
-		return false
-	}
-
 	// If the listener has no routes, create a new slice of routes with the given route.
 	if l.Routes == nil {
 		l.Routes = []ResourceReference{routeRef}
